Cleaned up entities
